# [kube-proxy: nftables] Optimize kube-proxy restart time