<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?
/kind feature
Performance improvement.

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

#### What this PR does / why we need it:

When running performance tests comparing iptables and nftables modes of kube-proxy, I noticed a significant difference for kube-proxy restart time. All the measurements depend on the server setup and may vary even on the same machine, but good enough to see the difference. Kernel version has a major effect on the results too. I am running an example setup with 10K svc, 2 EP each and measure the first transaction time after the kube-proxy restart. 

```
kernel 6.8.0-51-generic
iptables 
"SyncProxyRules complete" ipFamily="IPv4" elapsed="40.122804568s" 
nftables
"SyncProxyRules complete" ipFamily="IPv4" elapsed="3m14.380146902s" 
```
in the last year, thanks to Florian Westphal, some kernel optimizations were done to improve this difference, e.g. 
https://patchwork.ozlabs.org/project/netfilter-devel/patch/20240710085835.1957-1-fw@strlen.de/
https://patchwork.kernel.org/project/netdevbpf/patch/20240711093948.3816-3-pablo@netfilter.org/
and probably some more, which gives some improvement on the later kernel (the result are obtained on a different machine, only compare numbers within the same code block)
```
nftables
kernel 6.8.5-301.fc40.x86_64
"SyncProxyRules complete" ipFamily="IPv4" elapsed="5m33.410513986s"

kernel 6.12.11-100.fc40.x86_64
"SyncProxyRules complete" ipFamily="IPv4" elapsed="2m9.110746369s"
```
getting the newest kernel will take some time for users, and even with the newer kernel, nftables is still worse that iptables
```
kernel 6.12.11-100.fc40.x86_64
iptables
"SyncProxyRules complete" ipFamily="IPv4" elapsed="38.140467319s"
nftables
"SyncProxyRules complete" ipFamily="IPv4" elapsed="1m43.202901845s"
```
... Which means, we should try to optimize it!
One way to reduce the time the initial transaction takes is to not transact what is already there.
First commit does that for map/set elements and the second one does that for endpoint chains.
Local test results look like this:
```
kernel 6.8.0-51-generic
master
"SyncProxyRules complete" ipFamily="IPv4" elapsed="3m14.380146902s"
1st commit
"SyncProxyRules complete" ipFamily="IPv4" elapsed="2m4.707474487s"
1st+2nd commit
"SyncProxyRules complete" ipFamily="IPv4" elapsed="1m5.402122108s"
```
iptables result is
```
"SyncProxyRules complete" ipFamily="IPv4" elapsed="40.122804568s" 
```
so still not the same, but much closer.

There is another somewhat related problem with the time it takes to sync a new node in a cluster, which can't be optimized by just not transacting existing config and has to be solved in the kernel. My current results are
```
kernel 6.8.0-51-generic
nftables
"SyncProxyRules complete" ipFamily="IPv4" elapsed="1m46.926251367s"
iptables
"SyncProxyRules complete" ipFamily="IPv4" elapsed="34.158336568s"
```
```
kernel 6.12.11-100.fc40.x86_64
nftables
"SyncProxyRules complete" ipFamily="IPv4" elapsed="45.103989581s"
iptables
"SyncProxyRules complete" ipFamily="IPv4" elapsed="46.981954742s"
```
so newer kernel already helps here.

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

#### Special notes for your reviewer:
I will try to add some unit tests, when we finalize the code, because I am lazy to change unit tests on every iteration :)

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```